Use a single button for photo and video
