+++ app/src/main/res/drawable/ic_shutter.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_pressed="true" android:drawable="@drawable/ic_shutter_pressed" />
+    <item android:state_focused="true" android:drawable="@drawable/ic_shutter_focused" />
+    <item android:drawable="@drawable/ic_shutter_normal" />
+</selector>
