+++ app/src/main/res/layout/activity_main.xml
@@ -16,9 +16,8 @@
         android:id="@+id/settings_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:src="@android:drawable/ic_menu_slideshow"
-        android:layout_alignParentTop="true"
-        />
+        android:src="@android:drawable/ic_menu_manage"
+        android:layout_alignParentTop="true" />
 
     <ImageButton
         android:id="@+id/camera_button"
@@ -26,8 +25,7 @@
         android:layout_height="wrap_content"
         android:src="@android:drawable/ic_menu_camera"
         android:layout_alignParentTop="true"
-        android:layout_centerInParent="true"
-        />
+        android:layout_centerInParent="true" />
 
     <ImageButton
         android:id="@+id/mute_button"
@@ -37,32 +35,23 @@
         android:layout_toRightOf="@id/camera_button"
         android:src="@android:drawable/ic_btn_speak_now" />
 
-    <Button
-        android:id="@+id/image_capture_button"
-        android:layout_width="110dp"
-        android:layout_height="110dp"
+    <ImageButton
+        android:id="@+id/photo_button"
+        android:scaleType="fitCenter"
+        android:background="@drawable/ic_shutter"
+        android:layout_width="90dp"
+        android:layout_height="90dp"
         android:layout_marginBottom="20dp"
         android:layout_alignParentBottom="true"
-        android:elevation="2dp"
-        android:text="@string/take_photo" />
-
-    <Button
-        android:id="@+id/video_capture_button"
-        android:layout_width="110dp"
-        android:layout_height="110dp"
-        android:layout_marginBottom="20dp"
-        android:layout_alignParentBottom="true"
-        android:layout_marginLeft="20dp"
-        android:layout_toEndOf="@id/image_capture_button"
-        android:elevation="2dp"
-        android:text="@string/start_capture" />
+        android:layout_centerInParent="true"
+        android:elevation="2dp" />
 
     <ImageButton
         android:id="@+id/play_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
-        android:layout_toRightOf="@id/video_capture_button"
+        android:layout_toRightOf="@id/photo_button"
         android:src="@android:drawable/ic_media_pause"
         android:visibility="invisible"  />
