+++ app/src/main/res/values-b+it/strings.xml
@@ -1,8 +1,5 @@
 <resources>
     <string name="app_name">CamerArt</string>
-    <string name="take_photo">Fai Foto</string>
-    <string name="start_capture">Inizia Video</string>
-    <string name="stop_capture">Stoppa Video</string>
 
     <string name="capture_value_latency">latenza</string>
     <string name="capture_value_quality">qualità</string>
