+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,5 @@
 <resources>
     <string name="app_name">CamerArt</string>
-    <string name="take_photo">Take Photo</string>
-    <string name="start_capture">Start Capture</string>
-    <string name="stop_capture">Stop Capture</string>
 
     <string name="capture_value_latency">latency</string>
     <string name="capture_value_quality">quality</string>
